add: add adaptive footer, fix: fix sign-in link path
